backend/requirements.txt
absl-py==0.12.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
blinker==1.8.2
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-cors==4.0.0
flatbuffers==1.12
fonttools==4.56.0
fpdf2==2.8.2
frozenlist==1.5.0
gast==0.3.3
git-filter-repo==2.47.0
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.1.5
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==24.2
pathspec==0.12.1
Pillow==8.1.2
platformdirs==4.3.6
propcache==0.2.0
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.12.1
PyJWT==2.9.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tomli==2.2.1
twilio==9.4.5
typing_extensions==4.12.2
urllib3==1.26.3
Werkzeug==3.0.6
wrapt==1.12.1
yarl==1.15.2
zipp==3.4.1